[spirv-reader][ir] Handle propagation through multiple exits.
It's possible to setup a block where a value which needs to be
propagated dominates multiple blocks which jump to a continuing block.
In this case, we need to pass the propagated value through all of the
exits instead of just a single exit.

Change-Id: I8e4c37f37340899159a3d838bb2f38287efa924f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/245334
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: James Price <jrprice@google.com>
Auto-Submit: dan sinclair <dsinclair@chromium.org>