[spirv-reader][shader-io] Fold away lets to var's.
If there is a `let` of a `var` (which can happen from an `OpCopyObject`)
then we need to fold away the let and use the var directly in ShaderIO.
The var then folds away and becomes an input parameter.

Add support for folding through `lets`.

Bug: 42250952
Change-Id: I91db43a07cafcb4d1f1d9b50019316883e36db79
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/245594
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>