[spirv-reader][ir] Make sure `position` is used.
In SPIR-V, it's possible to create a `position` var, set it into the
`OpEntryPoint` interface, and then never store to `position`. This will
produce undefined values into the position.

In IR, we _must_ reference the `position` in order for it to be
considered part of the interface for the vertex shader. So, if we detect
that `position` is setup in `OpEntryPoint`, but never referenced, we'll
inject a store of `(0, 0, 0, 0)` into the `position` variable to satisfy
the interface rules.

Bug: 42250952
Change-Id: Ic37a4d6e34842bfd305dca22380df7527c469bed
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/245615
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Auto-Submit: dan sinclair <dsinclair@chromium.org>