[ir_to_program] Cleanup IRToProgram.
This Cl simplifies the binding tracking to just store the `ir_expr`,
`ast_expr` or `name` we care about and drops the use of the `variant`
and visitor code. The `SCOPED_NESTING` is removed in favour of the more
commonly used `TINT_SCOPED_ASSIGNMENT`.

Bug: 422532802
Change-Id: I6b3aa199276c26069f229b4fa6f17352859ec280
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/245614
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>