[vulkan] Add lazy memory tracking for ResourceMemAllocatorVk
While investigating high GPU memory usage for Graphite/Dawn/Vulkan
we found that lazily allocated memory in both Ganesh/Vulkan
and Graphite/Dawn/Vulkan are being tracked as part of total and used
vulkan memory allocations. Since presumably Graphite has more memory
that is lazily allocated it is better to track this separately.

Hopefully these memory allocations never happen, like on tiling GPUs,
but if an allocation is needed then hopefully it will be transient and
only exists for the duration of the draw.

This change tracks lazy allocated memory for dawn/vulkan separately.
For this we update AllocatorMemoryInfo and AllocationInfo for reporting
and tracking in ResourceMemoryAllocatorVk.

Bug: chromium:407730048
Change-Id: Idda25b1bc0d66ac2ab40e7f6545327bf079cbac8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/243655
Reviewed-by: Kyle Charbonneau <kylechar@google.com>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>