[spirv-reader][ir] Fix propagation when jumping if controls.
It's possible to jump the exit of an `if` and go straight to the exit of
a `loop` (a `break`). In that case, if we're propagating something from
inside the `if` we do not pass the result up through the `if` but
instead pass it directly through the `exit_loop` and return from the
`loop`. This is done by checking if the control has any exits before
attaching the propagated result.

Bug: 422539666
Change-Id: I03484a38c2758c7d39c0fe2310c649042e6e40b6
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/245757
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>