[wgsl] Support `subgroupMatrixLoad` and `subgroupMatrixStore` with `i8` and `u8`
This patch adds the overloads of `subgroupMatrixLoad` and
`subgroupMatrixStore` with `i8` and `u8` as the element types of the
subgroup matrices.

This patch also fixes the template of the subgroup matrix tests about
`i8` and `u8` by replacing all `array<i8>` with `array<i32>` and
replacing `array<u8>` with `array<u32>`.

Bug: 348702031, 403609083
Test: tint_unittests
Change-Id: Id41ddb92b294ab2ce80fd74a533031883868cc45
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/245454
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>
Reviewed-by: James Price <jrprice@google.com>