[spirv-reader][ir] Support `OpSpecConstantComposite`.
The `OpSpecConstantComposite` is the only way to build a composite
object from spec constants. In WGSL there is no equivalent, so if the
`OpSpecConstantComposite` has an attached `SpecId`, we error. In the
case there is no `SpecId` associated, we can turn the
`OpSpecConstantComposite` into a `Construct`. This construct needs to be
emitted in the block of usage, as it can not appear in the root block.

Bug: 398008657
Change-Id: I0ab4e55957aab63597d4e4eba87382738162422f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/245994
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>