[ir] When raising to a program handle unreachable as function terminator.
If an unreachable is the terminator of a function, and the function is
non-void, we swap the terminator out for a `return` of the function
return types zero value. This works around an issue where WGSL requires
a terminator, but does not have the behaviour analysis to determine if
the end of function is actually reachable or not.

Bug: 42250952
Change-Id: I5ef07147632d3ac54fdcc911cc844b959861626b
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/246075
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>