[dawn][emscripten] Fix ComboLimits compilation on Emscripten
A single copy of ComboLimits is used for both Dawn and Emscripten
targets, but it's generated with Dawn tags. Rather than try to generate
multiple copies of it (and move it to some other path), generate ifdefs
to make it work correctly on both (and switch it to params_all since
it's actually for all targets, not Dawn).

Bug: none
Change-Id: I7344c792a11117a715a5b9941ec8dbb27c726077
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/245240
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Auto-Submit: Kai Ninomiya <kainino@chromium.org>