[ir][validator] Fix return value check for void functions.
If a function returns `void` the `CheckReturn` in the IR validator was
just checking if the return `Value` was true. In the case where the
return has a value of `undef` then that check would not trigger as
`nullptr` is false.

Add a `HasValue` into the `ir::Return` so we can determine if anything
exists in the operands, instead of checking the truthyness of the value
itself.

Bug: 422810448
Change-Id: I9d0c72632144289fd05c2aecef82972ade757066
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/246154
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>