[spirv-reader][ir] Ignore continuing blocks without inbound branches.
If there are no branches into the continuing block, we can skip emitting
it. This works around an issue in SPIR-V where the continuing with no
in-bound branches can reference IDs from any of the blocks above it.
This makes it really hard to propagate the value up to the continuing
block.

Bug: 42250952
Change-Id: If855dbb4a743a93ea0adc3b72dd87bbc1cef8481
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/246274
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>