[spirv-reader][ir] Make sure outputs vars are referenced.
For each output var in the entry point interface, add a phony to the var
so it becomes part of the interface. The phony will then be removed by
the shader_io transform as any needed items will be moved into the io
structures.

Bug: 42250952
Change-Id: I97e0d83bf43ac3627bacd37d1d298ffe75f4824c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/246416
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>