[infra] Add a check for incompatible changes to ir.proto
This check is not designed to be all encompassing of any possible
incompatible change, just catching enumerated values changing in the
proto, and existing types being removed.

The parsing in the change is simple, not importing and using the proto
library, so there is lots of corner cases that will probably cause it to
bail, but it works for the existing ir.proto file and should work
reasonably for future versions that are well formed and have a readable
style. It not handling tricky 'technically' valid syntax is considered a
feature not a bug.

This check also hardcodes the location of the ir.proto file. This path
isn't expected to change, and the check can be updated if it does need
to move.

Fixes: 422414025

Change-Id: Ib4f51b21f90913429fe7af58d62389cc6e3e4768
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/247174
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>