[spirv-reader][ir] Handle PointSize applied to an `OpVariable`.
We previously handled `PointSize` when applied to a structure member,
but not when applied to an `OpVariable`. In this case, we can just not
emit the member into the output structure (because WGSL has no
PointSize) and set the initializer to `1.0f` to emulate the value for
any usage internal to the shader.

Bug: 42250952
Change-Id: Icc5e6a6e22ceac0156e13c7a79e6161de1a7cf84
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/246754
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>