[dawn] Improve E2E test for subgroup matrices
Fill the input matrices with different values so that they will detect
more problems in the implementations and drivers.

Use values that should not cause precision issues for small matrix
multiplies. Use a reference implementation to generate the expected
results on the CPU.

Refactor some of the code out to helper functions and classes so that
they can be reused for other tests in the future.

Fixed: 403609923
Change-Id: I4f266084396404057a7ac5f694ffbedc6ba423b4
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/245837
Reviewed-by: David Neto <dneto@google.com>
Commit-Queue: James Price <jrprice@google.com>