[tint][ir] Validate that discard is not directly in a non-fragment
The existing check correctly identities if a discard occurs in a
function that is called in a non-fragment entry point, but does not
catch if the function for the entry point itself contains a discard.
This is rooted to the fact that entry point functions are not considered
to be contained in any other functions, not even reflexively themselves.

This is fixed here by adding an explicit check in the validation rule
for the containing function being an endpoint. The alternate would be to
make a function be in its own set of containing functions, but that may
have wider ranging impact that would need further testing to avoid
infinite loops, since currently containing function is synonymous with
calling function.

Fixes: 424227913
Change-Id: If847c464a78db1e0ea2b90d32a2e52cdf82ac433
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/247334
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>