[spirv-reader][ir] Split input structs apart.
When a struct is provided as part of an input `var`, always split the
struct apart into entry point params. This fixes an issue when the
struct is shared with an output var, we don't know when to remove the
attributes on the struct.

Bug: 42250952
Change-Id: If4b9016857a34708c71356ec0fdfcddf4449c92a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/247434
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>