[tint] Basic scalarization support (max min clamp)
Various platforms seem to have issues with vectorized max,min, and
clamp builtins. For adreno this likely is due to a poor front end
implementation of handling nans.

For Intel (gen9) it is less clear where the implementation goes wrong
but what is clear is that scalarization resolves the issue.

Removed spirv only scalarization as it is duplicate. End2End test
still remains and passes.

Bug: 422144514, 407109052
Change-Id: Ie16617136be6e5993d5cf9c813eac2aa6b5492af
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/246095
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Peter McNeeley <petermcneeley@google.com>