[spirv-reader] Remove infinite loop test.
Remove test which generates an infinite loop. Both branches of the loop
are marked as `unreachable` this means we end up generating an infinite
loop as there is no break.

Just remove the test case, if this is provided as input, then the input
should be fixed to remove the infinite loop.

Bug: 42250952
Change-Id: Ib0452530ff6a58860580b6e1cc97302cb0a48c8e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/247495
Commit-Queue: David Neto <dneto@google.com>
Reviewed-by: David Neto <dneto@google.com>