[spirv-reader][ir] Guard against unhandled decorations for types.
In the type creation, assert that we've handled all of the provided
decorations. Currently they're just skipped, so it's hard to tell if
we're missing support. None of the decorations are handled yet, just an
assertion that we don't see any we haven't handled.

Bug: 391485267
Change-Id: I2fb152bc8a85319ea61c513ae1f853dcfb78474c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/247694
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>