[spirv-reader][ir] Handle `BufferBlock` annotation.
Support the `BufferBlock` annotation by re-writing the types of any
pointer of with a `BufferBlock` struct internal to it. This is done in
the parser as otherwise the emitted IR can contain things like runtime
arrays in uniform buffers which is invalid.

Bug: 391482883
Change-Id: I63e0f6a83943c6b53d9f0efeaf8d2ca03b8e6a6a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/247874
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: James Price <jrprice@google.com>