[dawn][opengl] Simplify GenerateCombinedSamplerInfo
The GL backend needs to translate runtime combining of the textures and
samplers in the shader to OpenGL's concept of combined textures and
samplers. During shader compilation, reflection data from Tint is used
to produce metadata related to that both for Dawn and Tint.

The metadata for Tint tells the GLSL writer what to name the combined
texture+sampler bindings that it generates. These names are then used to
query the resulting GL uniform location and the metadata for Dawn is
used to relate individual WebGPU texture or sampler bindings to a list
of GL texture units they affect.

GenerateCombinedSamplerInfo was very complicated as it went through
multiple rewrites as surrounding code was modified. Simplify it by:

 - Factoring the logic ran when a new combined texture and sampler is determined.
 - Making the difference of handling between external texture and regular texture more clear.
 - Avoiding the use of the "placeholder" sampler in CombinedSampler and using an std::optional instead.
   - Fixed a preexisting bug in the implementation of operator < for that structure.
 - Make it more clear what parameters are inputs and outputs.
 - Removing the redundant needsPlaceholderSampler output.

Also simplified the code in PipelineGL a bit since we only care about
the total set of all CombinedSampler and not the per-stage sets.

Bug: 411460085
Change-Id: Ic9b36aeaf8680deff5166efc66faa43090a53462
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/247035
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>