Mark WebGPU-WebGPU canvas copy tests as flaky on Nvidia Win
https://crrev.com/c/6464540 was an optimization for avoiding an extra
copy when using a WebGPU compatible shared image as the source image in
copyExternalImage. It was reverted due to a flaky crash in the D3D12
debug layer on Nvidia which is likely not a Chromium/Dawn bug and needs
to be investigated separately.

Mark these tests as flaky in preparation for relanding the CL.

Bug: chromium:412395829
Change-Id: I6a6a6964528d1f98b0467719f712430158f3fcf3
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/248114
Auto-Submit: Sunny Sachanandani <sunnyps@chromium.org>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>