[tint][glsl] Directly create all replacements in texture_polyfill
glsl.TexturePolyfill does several things one of which is introducing
GLSL's CombinedTextureSamplerVar for all combinations of texture+sampler
pairs statically used together in the shader (as well as for sampled
textures used without a sampler).

The logic used to create the CombinedTextureSamplerVar for
texture+sampler pairs as a pre-processing step but then would create the
ones for texture usages without samplers on demand, so that it could
reuse one of the texture+sampler replacements as an optimization.

Instead do the computation of the replacements for texture usages
without samplers in a second step of the preprocess, simplifying when
the various pieces of logic in that transform run. This will make it
easier to introduce support for binding_array in the transform in a
follow-up CL.

Bug: 411573957
Change-Id: I5074d8407e3bfd2649ebe963c18650d660843051
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/247395
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: James Price <jrprice@google.com>