[hlsl] Use IR by default for CLI and E2E tests.
The `hlsl-dxc-ir` and `hlsl-fxc-ir` formats are now skipped in the E2E
test runner. They will be removed when the CQ has been updated to no
longer use them.

The `hlsl-dxc` and `hlsl-fxc` expectations are currently set to the `ir`
outputs until the expectations are moved around.

Bug: 421916945
Change-Id: I1773b888b16abefe2505139676866d22cea08470
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/248314
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>