[tint][glsl] Support binding_array in texture_polyfill
Replaces the concept of finding the root variable for a given
texture/sampler value with the "path" which is a combination of a
variable and optionally an index to access that variable with.

Adds unittests for various kinds of paths for a texture value coming
from a binding array as well as tests showing that the whole
binding_array is duplicated in the bindings when used by multiple
samplers.

Bug: 411573957
Change-Id: I7cc718c03eeb410c84b0cdc180092543ccc81634
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/247396
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>