[spirv] Handle `stride` in `subgroupMatrixLoad` for `i8` `u8` subgroup matrices
In WGSL the `stride` means the number of subgroup matrix elements, while in
SPIR-V the `stride` is related to the type of the input pointer. So when
calling `subgroupMatrixLoad` to load an `i8` or `u8` subgroup matrix from an
`i32` or `u32` array, we need to divide 4 on the WGSL `stride` to turn it into
the SPIR-V `stride`.

Bug: 348702031, 403609083
Change-Id: I045ae20d39351dbd6aaed50581730dc228f62485
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/247254
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>