[spirv] Handle `offset` in `subgroupMatrixLoad` for `i8` `u8` matrices
In WGSL the `offset` means the number of subgroup matrix elements, while
in SPIR-V the `offset` is applied to the input pointer. So when calling
`subgroupMatrixLoad` to load an `i8` or `u8` subgroup matrix from an
`i32` or `u32` array, we need to divide 4 on the WGSL `offset` to match
the type of the SPIR-V input pointer.

No-Try because linux-dawn-rel is broken but this should be safe and
passed other bots.

No-Try: true
Bug: 348702031, 403609083
Test: tint_unittests
Change-Id: I52a2d89b6b041178e5a4e86e604d16ca2f3882cf
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/249834
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>