[tint][glsl] Handle binding_array in the unsampled texture_polyfill case.
The computation of the replacement for texture bindings that are
unsampled is different than for the sampled case because depth_texture
needs to be turned into texture_2d<f32>. When adding support for
binding_array this code path wasn't handled, incorrectly transforming
the IR when a texture only had unsampled usages.

Adds a test for the unsampled usage of the binding_array textures to
fix missing coverage.

Bug: 411573957
Change-Id: Ie531cc83a0d056e82d57136289eac9392330676c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/249215
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: James Price <jrprice@google.com>