[dawn] Fix nullability of WGPUPipelineLayoutDescriptor.bindGroupLayouts
The 'optional': true on this member was setting it WGPU_NULLABLE, which
is technically true, but we don't do this for any of the other array
members. The 'optional' here actually meant something very different,
that the *inner* object was nullable, not that the array was nullable.
So renamed that to 'array_element_optional'.

Bug: 427657664
Change-Id: I92bfb4f7694d472ddd1a1d7ebe31aada360d003e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/249396
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>