[spirv] Handle `offset` in `subgroupMatrixStore` for `i8` `u8` matrices
In WGSL the `offset` means the number of subgroup matrix elements, while
in SPIR-V the `offset` is applied to the output pointer. So when calling
`subgroupMatrixStore` to store an `i8` or `u8` subgroup matrix into an
`i32` or `u32` array, we need to divide 4 on the WGSL `offset` to match
the type of the SPIR-V output pointer.

Bug: 348702031, 403609083
Test: tint_unittests, dawn_end2end_tests
Change-Id: I4bc95caef59a0b5537d4c91a505128d0db89d0fe
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/249815
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Jiawei Shao <jiawei.shao@intel.com>