[doc] Update documentation about the build recipe location
Where the build recipes are stored has changed, so at a minimum the
links need to be updated. Also updated the text to be more
representative of the current state of the recipes, since there has
been a number of features/tasks added that are not in the original
description.

Change-Id: I82e5043143e212b5f266946388ac27e5d2971877
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/251154
Reviewed-by: James Price <jrprice@google.com>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>