[spirv-reader][ir] Add OpSpecConstantComposite id's to the id stack.
When creating a constant composite we were replacing the id meaning, but
not inserting the ID into the scope stack, so it did not appear to be
in scope when subsequently used.

Use the `AddValue` helper instead of inserting the value directly so
it's added to the id scope stack.

Bug: 429651458
Change-Id: I77db090d5f2fe59aa92ce31818cb679369ea99d9
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/250876
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: David Neto <dneto@google.com>
Reviewed-by: James Price <jrprice@google.com>