[spirv-reader][ir] Fix incorrect switch/loop exits.
In the case of an `if` where one of the branches jumps to a switch/loop
merge, and the other branch goes to the if merge which also jumps to the
switch/loop merge, from an IR perspective those two SPIR-V blocks would
refer to the same IR block. In this case, we need to differentiate by
looking at the `BranchConditional` and determining if we need to handle
this special terminator decision.

Bug: 42250952
Change-Id: I19d289bc455d7714b62898afc88ea16bd013a6fa
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/251197
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>