[tint] Abs polyfill for all backends
Now that the signed arithmetic polyfill has landed [0] the workaround
for abs (as max(x,-x)) will now function correctly.

The issue with abs(i32) was only proven to be observed on intel and
nvidia but have have chosen to proactively extend this fix to all
backends and devices. This decisions was made based on the polyfill
being very low overhead and assuming a commonality between all
backend compilers (LLVM).

[0]
https://dawn-review.googlesource.com/c/dawn/+/249135


Bug: 426999765
Change-Id: I2597cbcaacf8e68f93eaa7fb14d67d16cadcf970
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/247896
Commit-Queue: Peter McNeeley <petermcneeley@google.com>
Reviewed-by: James Price <jrprice@google.com>