[spirv-reader][ir] Build the texture worklist after replacing functions.
When processing the user call instructions to replace texture variables
we may fork/delete functions. But, if the instruction we delete is one
that we were going to replace the texture builtin call, we'll get
invalid arguments.

So, process all the variable changes, and then collect up the builtins
to process.

Bug: 428007402
Change-Id: I6b9f66552fbd06ae7535d9386d2054dcdf9da457
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/249137
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>