[dawn] Enable TINT_BUILD_SPV_READER in Wasm
Fix a bug where TINT_BUILD_SPV_READER would be disabled in builds that
had dawn_wasm, even for non-Wasm targets. (One GN build can contain both
Wasm and non-Wasm targets.)

We could probably just fix it to use is_wasm instead of dawn_wasm, but
there should be no reason to disable the SPIR-V Reader in Wasm builds.
(In fact, we definitely should support it so that Tint can be chained
with an HLSL or GLSL compiler in Wasm.)

CreateShaderModuleFromASM didn't work in Wasm because it depends on
wgpu::DawnShaderModuleSPIRVOptionsDescriptor. It's only used in one test
file, so move it to that test file unmodified.

Bug: none, 419596995
Change-Id: I20552d0086f8b1681afcec111fd75481100125aa
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/251775
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>