Fix ApiObjectList::Destroy() race
There is a race if ApiObjectList::Destroy() is going to destroy an
object and another thread drops the last reference to the object at the
same time. The other thread won't run DestroyImpl() but it will call
`delete this` and delete the object resulting in a use-after-free.
Keeping the ApiObjectList lock during DestroyImpl() calls won't work due
to https://crbug.com/42240929.

Have ApiObjectList::Destroy() acquire a reference to all of the
objects. RefCounted::TryAddRef() is added to expose the try to add
reference functionality that WeakRefData was already using.

Bug: 396294899, 429112750
Change-Id: Ib0dcad57bccac87835c2c6d1cab3609cf0e53ee7
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/251196
Commit-Queue: Kyle Charbonneau <kylechar@google.com>
Reviewed-by: Brandon Jones <bajones@chromium.org>