[emdawnwebgpu] Fix wgpuQuerySetGetType
Fix wgpuQuerySetGetType by translating the string to int instead of
letting JS coerce it.

WebGPU.QueryType.indexOf (which is slightly slower but avoids needing a
string-to-int array) wasn't possible because WebGPU.QueryType is an
object instead of an array (because the enum is sparse), so this changes
it to be a real array.

It also reduces the code size of this and other enums that are dense but
start at 1, *and* may be slightly faster for int-to-string lookups.
(Emscripten's load-and-reserialize for compiling library members was
already removing the unnecessary `undefined` in arrays that contained
`undefined`, but this change makes the generator avoid that too.)

An alternative would be to store the query type in memory and implement
this in webgpu.cpp. This should reduce the overall code size but it
would not work for imported query sets. This is something we probably
should consider in the future (tracked in bug 377760848).

Add SpotTests.cpp for Emdawnwebgpu smoke/regression tests, and a test
for this (copied from the bug report).

Fixed: 429140225
Bug: 377760848
Change-Id: I41e639cf798b558cd0c330258e690b074c4b32a3
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/252254
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>