[emdawnwebgpu] Fix string-to-int tables with Closure, enable Closure in tests
Fix string-to-int tables by (1) making them library-level items that are
explicit dependencies (separate from $WebGPU) and (2) wrapping them to
prevent Emscripten from removing the quotation marks on the keys, that
signal to Closure that they shouldn't be minified.

This has the added benefit of Emscripten being able to eliminate tables
that are unused by the application.

Enable --closure=1 in Release builds, and add a test which explicitly
tests this. (The GetCompilationInfo tests also catch this.)

Fixed: 424836759
Bug: 377760848
Change-Id: Iaee6d18516a5d54a49869ba76ec975c12cc0cc60
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/252255
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>