[dawn][opengl] Remove the GL-specific handling of un/filtering samplers.
This is a likely a remnant of when the OpenGL backend was initially
created, when the WebGPU spec didn't have the concept of "non filtering
samplers". GL requires that u/int textures be used only with nearest
filtering so each sampler was split into a two, a filtering and the
respective non-filtering one. This limitation is now part of the WebGPU
validation so it is no longer necessary to handle it in the GL backend.

Also makes opengl::Sampler follow the usual pattern for ::Create (with
an Initialize method), and remove the unused and confusing
opengl::PipelineLayout::GetTextureUnitsUsed.

Bug: 411573957
Change-Id: Ie29d697f7a9f30c1431f750eb6e1adaee9d1235d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/252134
Reviewed-by: Stephen White <senorblanco@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>