[emdawnwebgpu] Change link test to make function calls to test code size
Real function calls (even with bogus arguments) force all of the code to
be included, and also be able to invoke LTO inlining.

A volatile hack is used to prevent any of the arguments to the function
from being constexpr so they can't be constant-folded.

Bug: 414858419
Change-Id: Ic8e8245567990a55d15eaf504e75345c7ffcb4a3
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/252316
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>