[spirv-reader][ir] Fix loop continuing Phi issue.
If we're propagating a value through a loop continue block, but the
continue is unreachable, then we would have skipped emitting the body
and the propagated value won't exist.

In this case, just propagate an undef value as we don't actually care
what the real value was.

Fixed: 429625248
Change-Id: If2a7269546c8c203a16c36c96b15bc21383b3aa0
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/252935
Reviewed-by: David Neto <dneto@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>