Kotlin: Make WebGPU container class API more Kotlin idiomatic
Replace nullable surface with a getter that will raise an exception if
called in a context where the surface is not present.

The only circumstances in which a surface is not present is if the
caller did not pass in an Android surface when the WebGPU creator was
called.

Test: ./gradlew connectedAndroidTest
Bug: 373835938
Change-Id: Iaeeb03440bfab04b612956dd8d898cad3d269820
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/253994
Auto-Submit: Jim Blackler <jimblackler@google.com>
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Jim Blackler <jimblackler@google.com>