[dawn][opengl] Add typed integer for GL-specific indices.
The OpenGL backend has a few more kinds of indices used in the zoo of
indices for bindings: a flattened binding index across a whole pipeline
layout and the GL texture units. Add typed integer for them so it's
clearer what an index is instead of just GLuint.

Bug:
Change-Id: Id9f26598cf05059d4157b7ce81f5d0a6c5717a74
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/253495
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Stephen White <senorblanco@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>