Make AsyncWorkerThreadPool into an actual pool
Previously AsyncWorkerThreadPool would spawn a new thread per task,
which doesn't scale as we look at depending more on the task system
to handle a variety of smaller tasks off the main thread.

This change makes it into an actual thread pool of a fixed size to
reduce the per-task overhead.

Note that AsyncWorkerThreadPool is only used in applications that choose
not to override DawnPlatform::CreateWorkerTaskPool, such as the Dawn
tests. When used in Chromium the browser supplies it's own
WorkerTaskPool.

Bug: 430452846
Change-Id: I14d9d8aabd340732d0ab187b3f3e08efeab08a17
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/252975
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>