[dawn][opengl] Correctly allocate builtin metadata between stages.
Some texture builtins that query metadata need to be emulated by passing
that data in a UBO. Multiple stages may used such metadata which caused
issues with the previous approach to allocating an offset in the UBO for
the metadata. Replace it with a "registrar" that reuses the same offset
if possible.

Also message the datastructure storing the Dawn-side information of what
metadata needs to be put where to make it subjectively a bit nicer, and
use a FlatBindingIndex instead of BindGroupIndex+BindingIndex to address
it.

Bug: 427409135
Change-Id: I1326434517cadc782c683bfdd60aeae1fe9c1992
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/253496
Reviewed-by: Stephen White <senorblanco@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>