[dawn][opengl] Add for binding_array<texture> builtin emulation.
The emulation of builtins like textureNumLevels in the GL backend use a
UBO to contain the metadata of the texture. Support for binding_array
needs to be added by registering in Dawn that all elements of the array
need their data added to the UBO.

And add a test.

Bug: 427409135
Change-Id: I294795573c2baa061d574c406da5b5019582aa2a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/253734
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Stephen White <senorblanco@chromium.org>