[wgsl] Add chromium_print language feature
The idea is that this will always be advertised as an available
language feature, but might turn into a no-op (with a warning) if the
device cannot support printing for a given pipeline.

This makes it easy for developers to use print, without having to
enable an extension on the adapter and at the top of a shader.

Bug: 433534277
Change-Id: I60a8bca4447068a88d94f7c6d6ec0143d89b7962
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/254575
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>