[github] Use CMake to fetch deps in Emdawnwebgpu release jobs
Replace (most of) the custom dependency fetching in
package-emdawnwebgpu.sh in favor of DAWN_FETCH_DEPENDENCIES (which also
does a shallow fetch). Fetching the other dependencies surprisingly only
took an extra 22 seconds when tested on GitHub Actions, and it will be
needed if we start publishing prebuilt native binaries anyway.

This should fix the release jobs on GitHub Actions which currently fail
due to a missing Tint dependency on protobuf. (We don't actually need to
build Tint on that job, but we don't have a CMake option to disable
building Tint, and we don't need one if we just fetch all the deps.)

No-Try: true
Bug: 434707450
Change-Id: I49d4a7eb89e4cab638e0b801738055c341e9145c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/255115
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>