[tint][fuzz] Support generating IR corpus on the fly
Changes the behaviour of `fuzz -ir` to generate the IR corpus on the
fly if the default, WGSL, corpus is being used. Also requires -out
being set when -generate is used, because otherwise it just writes to
a tmp dir that is deleted on exit.

Bug: 348698576, 431997412
Change-Id: Iad92d43e85bfa0f606f9ba840569f4e286f1a973
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/255354
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>