[tint][fuzz] Implement corpus generation in Go
This implementation is planned to replace the existing standalone
Python script. That script is retained for now, since the build
recipes depend on it. Once this CL has landed and rolled, the build
recipes will be updated, and a follow on CL can remove the no longer
needed Python version.

Also includes a few incidental cleanups, including moving all of the
TODOs related to testing to the top of the file, since there is
multiple different issues that need to be resolved to be able to
implement testing coverage in this file, and I was tired of trying to
keep the TODOs attached to specific functions.

Bug: 348698576, 431997412

Change-Id: Ia5eb24d588cb1ce0b7bb5cecf7850b28f7b0d658
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/255434
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>