[tint] Add ICE callback mechanism
This enables all of the TINT_* macros that produce ICEs to accept an
optional parameter that provides a callback. Without this argument
they behave the same as they do today.

Add unit tests for the variants that use callbacks, and improve the
existing tests to ensure that the errors contain the source line
information.

Bug: 434966794
Change-Id: I53b6a8dee3f53d16f2074f94205cbb0886e549aa
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/255299
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>